+++ lesson7.py
@@ -25,149 +25,8 @@ Get name
 '''
 
 
-from tkinter import *
-from tkinter import messagebox as mb
-from tkinter import Tk, Frame, Menu
-
-
-class PhoneBook:
-    def __init__(self, name, number):
-        ''' init variables '''
-        self._name = name
-        self._number = number
-
-    def get_item(self):
-        ''' return item '''
-        return self._name, self._number
-
-    def set_item(self, name, number):
-        ''' set item '''
-        self._name = name
-        self._number = number
-
-
-class frame_window(Frame):
-    dict_phones = {}
-    list_of_phonebook = []
-
-    def __init__(self):
-        super().__init__()
-        self.entry_name = Entry()
-        self.entry_phone = Entry()
-        self.l_nom = Label(text="Name", font="Arial 10")
-        self.l_phone = Label(text="Phone", font="Arial 10")
-        self.text_go = Text(height=30, width=70, font='Arial 10', wrap=WORD)
-        self.main_menu = Menu(self.master)
-        self.file_menu = Menu(self.main_menu, tearoff=0)
-        self.dict_menu = Menu(self.main_menu, tearoff=0)
-        self.init_gui()
-
-    def init_gui(self):
-        self.entry_name.pack(pady=10)
-        self.entry_name.place(x=50, y=1)
-        self.entry_phone.pack(pady=10)
-        self.entry_phone.place(x=250, y=1)
-        self.l_nom.place(x=1, y=1)
-        self.l_phone.place(x=200, y=1)
-        self.text_go.pack()
-        self.text_go.place(x=1, y=30)
-        self.master.config(menu=self.main_menu)
-        self.dict_menu.add_command(label="Show all dict",
-                                   command=self.show_dict)
-        self.dict_menu.add_command(label="Add to dict",
-                                   command=self.add_to_dict)
-        self.dict_menu.add_command(label="Find by name",
-                                   command=self.find_by_name)
-        self.dict_menu.add_command(label="Find by phone",
-                                   command=self.find_by_phone)
-        self.dict_menu.add_command(label="Find repetition",
-                                   command=self.find_repetition)
-        self.dict_menu.add_command(label="Delete item by name",
-                                   command=self.delete_name)
-        self.file_menu.add_command(label="Exit", command=self.exit_app)
-        self.main_menu.add_cascade(label="File", menu=self.file_menu)
-        self.main_menu.add_cascade(label="Dictionary", menu=self.dict_menu)
-
-    def exit_app(self):
-        quit()
-
-    def show_dict(self):
-        ''' output all phonebook to display '''
-        self.text_go.delete('1.0', END)
-        for line in self.list_of_phonebook:
-            name, phone = line.get_item()
-            str_output = "Name {} Phone {}\n".format(name, phone)
-            self.text_go.insert(1.0, str_output)
-
-    def delete_name(self):
-        ''' delete item of phonebook by name '''
-        self.text_go.delete('1.0', END)
-        for line in self.list_of_phonebook:
-            name, phone = line.get_item()
-            if name == self.entry_name.get():
-                self.list_of_phonebook.remove(line)
-                self.text_go.insert(1.0, 'Well done!')
-                self.entry_name.delete(0, END)
-                break
-        else:
-            self.text_go.insert(1.0, 'Nothing to delete')
-
-    def add_to_dict(self):
-        ''' add to phinebook new item '''
-        if not (self.entry_name.get() == '' or self.entry_phone.get() == ''):
-            item = PhoneBook(self.entry_name.get(), self.entry_phone.get())
-            self.list_of_phonebook.append(item)
-            self.entry_name.delete(0, END)
-            self.entry_phone.delete(0, END)
-        else:
-            mb.showerror("Error", "Enter Name and Phone before")
-
-    def find_by_name(self):
-        ''' search  by name '''
-        self.text_go.delete('1.0', END)
-        for line in self.list_of_phonebook:
-            name, phone = line.get_item()
-            if name == self.entry_name.get():
-                str_output = "Name {} Phone {}\n".format(name, phone)
-                self.text_go.insert(1.0, str_output)
-                self.entry_name.delete(0, END)
-                break
-        else:
-            self.text_go.insert(1.0, 'Nothing\n')
-
-    def find_by_phone(self):
-        ''' search  by number '''
-        self.text_go.delete('1.0', END)
-        for line in self.list_of_phonebook:
-            name, phone = line.get_item()
-            if phone == self.entry_phone.get():
-                str_output = "Name {} Phone {}\n".format(name, phone)
-                self.text_go.insert(1.0, str_output)
-                self.entry_name.delete(0, END)
-                break
-        else:
-            self.text_go.insert(1.0, 'Nothing\n')
-
-    def find_repetition(self):
-        self.text_go.delete('1.0', END)
-        total = 0
-        rep = 0
-        for line in self.list_of_phonebook:
-            name, phone = line.get_item()
-            x = ''
-            for i in phone:
-                if x == i:
-                    rep += 1
-                else:
-                    rep = 0
-                if rep == 2:
-                    total += 1
-                    rep = 0
-                    self.text_go.insert(1.0, 'Phone ' + str(phone)
-                                        + " Name " + name + '\n')
-                    break
-                x = i
-        self.text_go.insert(1.0, 'Total: ' + str(total) + '\n')
+from tkinter import Tk
+from visualisation_data.frame import FrameWindow
 
 
 def main():
@@ -175,7 +34,7 @@ def main():
     root.title("Dictionary")
     root.geometry('500x400')
     root.resizable(False, False)
-    app = frame_window()
+    app = FrameWindow()
     root.mainloop()
